[NVim] Improve the LSP configuration
As I am planning of using different LSP servers between public and work
configurations, I need a smart way having to register the servers that
might be available in some parts but not in others.

This change allows to reuse the registration function and attaches the
same set of lsp functions to all the servers.